requirements.txt
# Core dependencies for stroke classification web app
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
Pillow>=9.5.0
numpy>=1.24.0,<2.0
opencv-python-headless==4.8.1.78
pandas>=1.5.0

# Flask dependencies
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# For production deployment
requests>=2.31.0
gunicorn==21.2.0
psutil>=5.9.0
